requirements.txt
Flask>=2.0
python-dotenv>=0.15
requests>=2.25
networkx>=2.5
pandas>=1.2
numpy>=1.20
matplotlib>=3.3 
tqdm>=4.50
Flask-CORS>=3.0 